pyproject.toml
[project]
name = "pinecone-quickstart"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pinecone>=5.4.2",
    "python-dotenv>=1.0.1",
]
